fix(core): move printf declarations before corelib
