+++ core/tests/test_data/llvm/fib_main.ll
@@ -9,6 +9,24 @@ entry:
   ret i252 %res
 }
 
+declare i32 @printf(ptr, ...)
+
+define i32 @print_felt(i252 %0) {
+entry:
+  %prefix = alloca [5 x i8], align 1
+  store [5 x i8] c"%ld\0A\00", ptr %prefix, align 1
+  %printed_characters_n = call i32 (ptr, ...) @printf(ptr %prefix, i252 %0)
+  ret i32 %printed_characters_n
+}
+
+define i32 @print_double_felt(i503 %0) {
+entry:
+  %prefix = alloca [5 x i8], align 1
+  store [5 x i8] c"%ld\0A\00", ptr %prefix, align 1
+  %printed_characters_n = call i32 (ptr, ...) @printf(ptr %prefix, i503 %0)
+  ret i32 %printed_characters_n
+}
+
 define { i252, i252 } @"dup<felt>"(i252 %0) {
 entry:
   %res_ptr = alloca { i252, i252 }, align 8
@@ -61,24 +79,6 @@ entry:
 define i252 @"felt_const<30>"() {
 entry:
   ret i252 30
-}
-
-declare i32 @printf(ptr, ...)
-
-define i32 @print_felt(i252 %0) {
-entry:
-  %prefix = alloca [5 x i8], align 1
-  store [5 x i8] c"%ld\0A\00", ptr %prefix, align 1
-  %printed_characters_n = call i32 (ptr, ...) @printf(ptr %prefix, i252 %0)
-  ret i32 %printed_characters_n
-}
-
-define i32 @print_double_felt(i503 %0) {
-entry:
-  %prefix = alloca [5 x i8], align 1
-  store [5 x i8] c"%ld\0A\00", ptr %prefix, align 1
-  %printed_characters_n = call i32 (ptr, ...) @printf(ptr %prefix, i503 %0)
-  ret i32 %printed_characters_n
 }
 
 define { i252 } @"fib_caller::fib_caller::fib"(i252 %0, i252 %1, i252 %2) {
